LayoutTests/http/tests/notifications/legacy/request-expected.txt

This test checks that a request for permission is made.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.webkitNotifications.checkPermission() is 1
PASS window.webkitNotifications.checkPermission() is 0
PASS successfullyParsed is true

TEST COMPLETE
